tr: move TapTree into taptree module
This completes the move of all the taptree-related stuff into its own
module, where it will be easier to mess around with this stuff.
